requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.14
django-admin-tools==0.9.1
django-allauth==0.44.0
django-cors-headers==3.8.0
django-import-export==2.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
